Install spigot only once

+++ scripts/entrypoint.sh
@@ -5,13 +5,16 @@ echo "***** Entrypoint..."
 echo " -------------------- "
 
 # Install Spigot if required
-echo "***** Installing Spigot"
-JVM_MAX_MEMORY="Xmx${JAVA_MEMORY}M"
-cd /McMyAdmin/Minecraft/spigot/
-java -"$JVM_MAX_MEMORY" -jar /McMyAdmin/Minecraft/spigot/BuildTools.jar --rev "$MINECRAFT_VERSION" > spigot.log 2>&1
-cp /McMyAdmin/Minecraft/spigot/spigot-*.jar /McMyAdmin/Minecraft/
-mv /McMyAdmin/Minecraft/minecraft_server.jar /McMyAdmin/Minecraft/minecraft_server.jar_backup
-mv /McMyAdmin/Minecraft/spigot-*.jar /McMyAdmin/Minecraft/minecraft_server.jar
+if [ "$USE_SPIGOT" = true ] && [ ! -f /McMyAdmin/Minecraft/spigot/.buildSuccess ] ; then
+  echo "***** Installing Spigot"
+  JVM_MAX_MEMORY="Xmx${JAVA_MEMORY}M"
+  cd /McMyAdmin/Minecraft/spigot/
+  java -"$JVM_MAX_MEMORY" -jar /McMyAdmin/Minecraft/spigot/BuildTools.jar --rev "$MINECRAFT_VERSION" > spigot.log 2>&1
+  cp /McMyAdmin/Minecraft/spigot/spigot-*.jar /McMyAdmin/Minecraft/
+  mv /McMyAdmin/Minecraft/minecraft_server.jar /McMyAdmin/Minecraft/minecraft_server.jar_backup
+  mv /McMyAdmin/Minecraft/spigot-*.jar /McMyAdmin/Minecraft/minecraft_server.jar
+  touch /McMyAdmin/Minecraft/spigot/.buildSuccess
+fi
 
 # Configure McMyAdmin
 echo "***** Configuring McMyAdmin conf file"
